`slack_preset.slack_user_id`

+++ src/presets/delete.js
@@ -20,10 +20,24 @@ module.exports = async function slackPresetDelete(req, res) {
   }
 
   const db = await req.db();
+  const dbOauthResp = await db.query(sql`
+    SELECT s.id, s.user_id, s.access_token FROM slack_oauth s
+    WHERE s.id = ${req.body.slack_oauth_id} AND s.revoked = false
+    LIMIT 1
+  `);
+
+  const oauth = dbOauthResp.rows[0];
+
+  if (!oauth) {
+    res.statusCode = TODO_BAD_REQUEST;
+
+    return;
+  }
+
   await db.query(sql`
     DELETE FROM slack_preset
     WHERE id = ${req.body.id}
-      AND slack_oauth_id = ${req.body.slack_oauth_id}
+      AND slack_user_id = ${oauth.user_id}
   `);
 
   res.writeHead(303, {
